Some Modification to Cisco NEXT JS
Dynamic Attribute Change Of Link or LinkSet Used interval for node
color change. next.js is modefied. Because if nodeSet are used (this
enables the linksets also) when viewing the nodeSets and the update
comes about the node inside the nodeset you can not change the node
attributes as the nodedictionary and the linkdictionary are not include
the subnodes. I added tow more dictionaries that inculudes the all
nodes and links in the topology.
MynodeDictionary and MylinkDictionary are added to reach every
link,node etc while updating data. Current getLink, and getNode are
using nodeDictionary, linkDictionary which are using also inside
next.js. Current modification to them affect next.js. So I need to
define dictionaries that I can reach at any time. Also every link, node
and Nodeset are added to the dictionaries at
nx.data.ObservableGraph.Edges, nx.data.ObservableGraph.Links,
nx.data.ObservableGraph.EdgeSet classes (at the very end of the
_addEdge, _addEdgeSet and _addLink methods before return statement.

For nodeSet, nodeSet status can be calculated while loading the
topology add adding node's to nodeset. If not also nodeset class and
corresponding method should be added like linkSet